separate resolvers config
